requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
redis==5.2.1
pydantic-settings==2.7.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
apscheduler==3.11.0
